common functions to utils
