+++ tools/hejret_mirna_cnn.py
@@ -1,10 +1,9 @@
-import argparse
 import os
 import urllib.request 
 import pandas as pd
 from tensorflow import keras as k
 
-from mirbind import one_hot_encoding
+from utils import parse_args, one_hot_encoding
 
 def predict_probs(df, miRNA_col, gene_col, model):
     """
@@ -33,18 +32,7 @@ def get_model_path():
     return model_path
 
 if __name__ == '__main__':
-    # Parse command line arguments
-    parser = argparse.ArgumentParser(description='Hejret miRNA CNN prediction.')
-    # Path to the input file 
-    parser.add_argument('--input', type=str, help='Path to the input file - miRNA and a gene sequence in a tab-separated format.')
-    # Name of column containing miRNA sequences
-    parser.add_argument('--miRNA_column', type=str, help='Name of the column containing miRNA sequences')
-    # Name of column containing gene sequences
-    parser.add_argument('--gene_column', type=str, help='Name of the column containing gene sequences')
-    # Path to the output file
-    parser.add_argument('--output', type=str, help='Path to the output file')
-    # Parse the arguments
-    args = parser.parse_args()
+    args = parse_args('Hejret miRNA CNN')
 
     # Read the input file
     data = pd.read_csv(args.input, sep='\t')
