+++ tools/seed.py
@@ -1,5 +1,6 @@
-import argparse
 import pandas as pd
+
+from utils import parse_args
 
 def rev_compl(st):
     nn = {'A': 'T', 'C': 'G', 'G': 'C', 'T': 'A'}
@@ -68,18 +69,7 @@ def classify_site(gene, miRNA, get_seeds):
     return 0
 
 if __name__ == '__main__':
-    # Parse command line arguments
-    parser = argparse.ArgumentParser(description='Seed match')
-    # Path to the input file 
-    parser.add_argument('--input', type=str, help='Path to the input file - miRNA and a gene sequence in a tab-separated format.')
-    # Name of column containing miRNA sequences
-    parser.add_argument('--miRNA_column', type=str, help='Name of the column containing miRNA sequences')
-    # Name of column containing gene sequences
-    parser.add_argument('--gene_column', type=str, help='Name of the column containing gene sequences')
-    # Path to the output file
-    parser.add_argument('--output', type=str, help='Path to the output file')
-    # Parse the arguments
-    args = parser.parse_args()
+    args = parse_args('Seed')
 
     # Read the input file
     data = pd.read_csv(args.input, sep='\t')
